+++ docs/CHANGELOG.rst
@@ -19,6 +19,8 @@ Changed
 
 Fixed
 -----
+- Fix handling of missing post-peak QC reports in MultiQC
+  when MACS2 is run in broad peak mode
 
 
 ===================